# docs/swagger.yaml
host: 127.0.0.1:3000
info:
  contact:
    email: drh.dimasrestu@gmail.com
    name: Dimas Restu Hidayanto
    url: https://github.com/dimaskiddo
  description: This is WhatsApp Multi-Device Implementation in Go REST API
  title: Go WhatsApp Multi-Device REST API
  version: 1.x
paths:
  /api/v1/whatsapp:
    get:
      description: Get The Server Status
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: Show The Status of The Server
      tags:
      - Root
  /api/v1/whatsapp/auth:
    get:
      description: Get Authentication Token
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - BasicAuth: []
      summary: Generate Authentication Token
      tags:
      - Root
  /api/v1/whatsapp/group:
    get:
      description: Get Joined Groups Information from WhatsApp
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - BearerAuth: []
      summary: Get Joined Groups Information
      tags:
      - WhatsApp Group
  /api/v1/whatsapp/login:
    post:
      consumes:
      - multipart/form-data
      description: Get QR Code for WhatsApp Multi-Device Login
      parameters:
      - default: html
        description: Change Output Format in HTML or JSON
        enum:
        - html
        - json
        in: formData
        name: output
        type: string
      produces:
      - application/json
      - text/html
      responses:
        "200":
          description: ""
      security:
      - BearerAuth: []
      summary: Generate QR Code for WhatsApp Multi-Device Login
      tags:
      - WhatsApp Authentication
  /api/v1/whatsapp/logout:
    post:
      description: Make Device Logout from WhatsApp Multi-Device
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - BearerAuth: []
      summary: Logout Device from WhatsApp Multi-Device
      tags:
      - WhatsApp Authentication
  /api/v1/whatsapp/registered:
    get:
      description: Check WhatsApp Personal ID or Group ID is Registered
      parameters:
      - description: WhatsApp Personal ID or Group ID to Check
        in: query
        name: msisdn
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - BearerAuth: []
      summary: Check If WhatsApp Personal ID or Group ID is Registered
      tags:
      - WhatsApp Authentication
  /api/v1/whatsapp/send/audio:
    post:
      consumes:
      - multipart/form-data
      description: Send Audio Message to Spesific WhatsApp Personal ID or Group ID
      parameters:
      - description: Destination WhatsApp Personal ID or Group ID
        in: formData
        name: msisdn
        required: true
        type: string
      - description: Audio File
        in: formData
        name: audio
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - BearerAuth: []
      summary: Send Audio Message
      tags:
      - WhatsApp Message
  /api/v1/whatsapp/send/contact:
    post:
      consumes:
      - multipart/form-data
      description: Send Contact Message to Spesific WhatsApp Personal ID or Group
        ID
      parameters:
      - description: Destination WhatsApp Personal ID or Group ID
        in: formData
        name: msisdn
        required: true
        type: string
      - description: Contact Name
        in: formData
        name: name
        required: true
        type: string
      - description: Contact Phone
        in: formData
        name: phone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - BearerAuth: []
      summary: Send Contact Message
      tags:
      - WhatsApp Message
  /api/v1/whatsapp/send/document:
    post:
      consumes:
      - multipart/form-data
      description: Send Document Message to Spesific WhatsApp Personal ID or Group
        ID
      parameters:
      - description: Destination WhatsApp Personal ID or Group ID
        in: formData
        name: msisdn
        required: true
        type: string
      - description: Document File
        in: formData
        name: document
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - BearerAuth: []
      summary: Send Document Message
      tags:
      - WhatsApp Message
  /api/v1/whatsapp/send/image:
    post:
      consumes:
      - multipart/form-data
      description: Send Image Message to Spesific WhatsApp Personal ID or Group ID
      parameters:
      - description: Destination WhatsApp Personal ID or Group ID
        in: formData
        name: msisdn
        required: true
        type: string
      - description: Caption Image Message
        in: formData
        name: caption
        required: true
        type: string
      - description: Image File
        in: formData
        name: image
        required: true
        type: file
      - default: false
        description: Is View Once
        in: formData
        name: viewonce
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - BearerAuth: []
      summary: Send Image Message
      tags:
      - WhatsApp Message
  /api/v1/whatsapp/send/link:
    post:
      consumes:
      - multipart/form-data
      description: Send Link Message to Spesific WhatsApp Personal ID or Group ID
      parameters:
      - description: Destination WhatsApp Personal ID or Group ID
        in: formData
        name: msisdn
        required: true
        type: string
      - description: Link Caption
        in: formData
        name: caption
        type: string
      - description: Link URL
        in: formData
        name: url
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - BearerAuth: []
      summary: Send Link Message
      tags:
      - WhatsApp Message
  /api/v1/whatsapp/send/location:
    post:
      consumes:
      - multipart/form-data
      description: Send Location Message to Spesific WhatsApp Personal ID or Group
        ID
      parameters:
      - description: Destination WhatsApp Personal ID or Group ID
        in: formData
        name: msisdn
        required: true
        type: string
      - description: Location Latitude
        in: formData
        name: latitude
        required: true
        type: number
      - description: Location Longitude
        in: formData
        name: longitude
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - BearerAuth: []
      summary: Send Location Message
      tags:
      - WhatsApp Message
  /api/v1/whatsapp/send/sticker:
    post:
      consumes:
      - multipart/form-data
      description: Send Sticker Message to Spesific WhatsApp Personal ID or Group
        ID
      parameters:
      - description: Destination WhatsApp Personal ID or Group ID
        in: formData
        name: msisdn
        required: true
        type: string
      - description: Sticker File
        in: formData
        name: sticker
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - BearerAuth: []
      summary: Send Sticker Message
      tags:
      - WhatsApp Message
  /api/v1/whatsapp/send/text:
    post:
      consumes:
      - multipart/form-data
      description: Send Text Message to Spesific WhatsApp Personal ID or Group ID
      parameters:
      - description: Destination WhatsApp Personal ID or Group ID
        in: formData
        name: msisdn
        required: true
        type: string
      - description: Text Message
        in: formData
        name: message
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - BearerAuth: []
      summary: Send Text Message
      tags:
      - WhatsApp Message
  /api/v1/whatsapp/send/video:
    post:
      consumes:
      - multipart/form-data
      description: Send Video Message to Spesific WhatsApp Personal ID or Group ID
      parameters:
      - description: Destination WhatsApp Personal ID or Group ID
        in: formData
        name: msisdn
        required: true
        type: string
      - description: Caption Video Message
        in: formData
        name: caption
        required: true
        type: string
      - description: Video File
        in: formData
        name: video
        required: true
        type: file
      - default: false
        description: Is View Once
        in: formData
        name: viewonce
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - BearerAuth: []
      summary: Send Video Message
      tags:
      - WhatsApp Message
schemes:
- http
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"